+++ src/types/@blinkmobile/aws-s3.d.ts
@@ -1,1 +1,0 @@
-declare module '@blinkmobile/aws-s3'